+++ mobile/android/android-components/buildSrc/src/main/java/Gecko.kt
@@ -6,12 +6,12 @@ internal object GeckoVersions {
     /**
      * GeckoView Nightly Version.
      */
-    const val nightly_version = "75.0.20200309091841"
+    const val nightly_version = "76.0.20200310094445"
 
     /**
      * GeckoView Beta Version.
      */
-    const val beta_version = "74.0.20200227210932"
+    const val beta_version = "75.0.20200309155231"
 
     /**
      * GeckoView Release Version.